feat(pwa): disable home Play CTA when offline
Closes the v1 DoD item "challenge submission button disabled offline"
at the entry point — the Daily Guess button on the home page is now
disabled with a translated hint when the browser reports offline. The
existing top offline banner still surfaces from anywhere in the app.

Adds a small reusable \`useOnline()\` hook and refactors OfflineIndicator
to consume it (DRY).
