+++ packages/frontend/public/locales/en/translation.json
@@ -269,6 +269,7 @@
     "dailyGuess": "Daily Guess",
     "geoCta": "Geo Mode",
     "guestHint": "No account needed — play instantly as a guest",
+    "offlineHint": "You are offline — reconnect to play today's challenge",
     "previewBadge": "Today's challenge",
     "previewHeading": "Think you know this game?",
     "previewSubtitle": "Click to play the full 10-screenshot challenge.",
