+++ packages/frontend/public/locales/fr/translation.json
@@ -269,6 +269,7 @@
     "dailyGuess": "Défi Quotidien",
     "geoCta": "Mode Géo",
     "guestHint": "Aucun compte requis — jouez instantanément en invité",
+    "offlineHint": "Vous êtes hors ligne — reconnectez-vous pour jouer le défi du jour",
     "previewBadge": "Défi du jour",
     "previewHeading": "Pensez-vous reconnaître ce jeu ?",
     "previewSubtitle": "Cliquez pour jouer le défi complet de 10 captures.",
